Moving the code of quarkus console to a separate module
