+++ plugins/quarkus-backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@qshift/plugin-quarkus-backend",
-  "version": "0.1.18",
+  "version": "0.1.19",
   "main": "src/index.ts",
   "types": "src/index.ts",
   "license": "Apache-2.0",
