+++ plugins/quarkus-console/src/components/QuarkusComponent.tsx
