+++ plugins/quarkus-console/src/components/QuarkusConsole/QuarkusConsolePage.test.tsx
