+++ plugins/quarkus-console/src/components/QuarkusConsole/QuarkusConsolePage.tsx
