+++ plugins/quarkus-console/src/components/QuarkusConsole/index.ts
