+++ plugins/quarkus-console/src/components/QuarkusConsoleFetch/data.ts
