+++ plugins/quarkus-console/src/components/QuarkusConsoleFetch/type.ts
