+++ plugins/quarkus-console/src/components/QuarkusIcon/QuarkusIcon.tsx
