+++ plugins/quarkus-console/src/components/QuarkusIcon/index.ts
