+++ plugins/quarkus-console/src/index.ts
@@ -1,0 +1,4 @@
+export { QuarkusConsolePlugin, QuarkusConsolePage } from './plugin';
+export { QuarkusComponent } from './components/QuarkusComponent'
+export { QuarkusComponentPodsTable } from './components/QuarkusComponentPodsTable'
+export { QuarkusIcon } from './components/QuarkusIcon';
