+++ plugins/quarkus-console/src/models.ts
