+++ plugins/quarkus-console/src/plugin.test.ts
