+++ plugins/quarkus-console/src/routes.ts
