+++ plugins/quarkus-console/src/services/K8sResourcesContext.ts
