+++ plugins/quarkus-console/src/services/getKubernetesResources.ts
