+++ plugins/quarkus-console/src/services/useAllWatchResources.ts
