+++ plugins/quarkus-console/src/services/useK8sObjectsResponse.ts
