+++ plugins/quarkus-console/src/types/route.ts
