+++ plugins/quarkus-console/src/types/types.ts
