+++ plugins/quarkus-console/turbo.json
@@ -1,0 +1,9 @@
+{
+  "extends": ["//"],
+  "pipeline": {
+    "tsc": {
+      "outputs": ["../../dist-types/plugins/quarkus/**"],
+      "dependsOn": ["^tsc"]
+    }
+  }
+}
