+++ plugins/quarkus/src/index.ts
@@ -1,6 +1,2 @@
 export { QuarkusExtensionListField } from './plugin';
 export { QuarkusQuickstartPickerField } from './plugin';
-export { QuarkusConsolePlugin, QuarkusConsolePage } from './plugin';
-export { QuarkusComponent } from './components/QuarkusComponent'
-export { QuarkusComponentPodsTable } from './components/QuarkusComponentPodsTable'
-export { QuarkusIcon } from './components/QuarkusIcon';
